Refactor to suggested changes
